+++ R/install_dependencies.R
@@ -2,7 +2,8 @@
 #'
 #' Installs the necessary command-line tools ('blast' and 'efetch') into conda environments if they are not already available. This ensures that all dependencies required by the package are present and up-to-date.
 #'
-#' @param verbose A character string specifying the verbosity level during environment creation. Options include `"silent"`, `"output"`, `"full"` etc. Default is `"silent"`.
+#' @param verbose A character string specifying the verbosity level during environment creation.
+#'   Options include `"silent"`, `"cmd"`, `"output"`, or `"full"`. Default is `"silent"`.
 #' @param force A logical value indicating whether to force the re-creation of the Conda environments even if they already exist. Default is `FALSE`.
 #'
 #' @return Invisibly returns `TRUE` after attempting to install the dependencies.
@@ -13,13 +14,16 @@
 #' install_dependencies()
 #'
 #' # Install dependencies with verbose output
-#' install_dependencies(verbose = "verbose")
+#' install_dependencies(verbose = "output")
 #'
 #' # Force re-installation of dependencies
 #' install_dependencies(force = TRUE)
 #' }
 #' @export
-install_dependencies <- function(verbose = "silent", force = FALSE) {
+install_dependencies <- function(
+  verbose = c("silent", "cmd", "output", "full"),
+  force = FALSE
+) {
   check_cmd(
     cmd = "blastn",
     env_name = "blastr-blast-env",