+++ R/parallel_blast.R
@@ -1,28 +1,41 @@
-#' @title Run BLAST on parallel
+#' @title Run Parallelized BLAST
 #'
-#' @description Run parallel BLAST for set of sequences
+#' @description Run parallel BLAST for a set of sequences
 #'
 #' @inheritParams get_blast_results
 #'
 #' @param asvs Character vector with sequences
 #' @param out_file Complete path to output .csv file on an existing folder.
 #' @param out_RDS Complete path to output RDS file on an existing folder.
-#' @param total_cores Total available cores to run BLAST in parallel. Check your max with *future::availableCores()*
+#' @param total_cores Total available cores to run BLAST in parallel.
+#'   Check your max with *future::availableCores()*.
 #' @param blast_type BLAST+ executable to be used on search.
 #'
+#' @inheritParams rlang::args_dots_empty
+
 #' @return A tibble with the BLAST tabular output.
 #'
 #' @examples
 #' \dontrun{
-#' blast_res <- BLASTr::parallel_blast(
-#'   asvs = ASVs_test, # vector of sequences to be searched
-#'   db_path = "/data/databases/nt/nt", # path to a formatted blast database
+#' dna_fasta_path <- fs::path_package(
+#'   "BLASTr", "extdata", "minimal_db_blast",
+#'   ext = "fasta"
+#' )
+#' temp_db_path <- fs::path_temp("minimal_db_blast")
+#' make_blast_db(fasta_path = dna_fasta_path, db_path = temp_db_path)
+#' asvs_string <- c(
+#'   "CTAGCCATAAACTTAAATGAAGCTATACTAA",
+#'   "ACTCGTTCGCCAGAGTACTACAAGCGAAAG"
+#' )
+#' blast_res <- parallel_blast(
+#'   asvs = asvs_string, # vector of sequences to be searched
+#'   db_path = temp_db_path, # path to a formatted blast database
 #'   out_file = NULL, # path to a .csv file to be created with results (on an existing folder)
 #'   out_RDS = NULL, # path to a .RDS file to be created with results (on an existing folder)
 #'   perc_id = 80, # minimum identity percentage cutoff
+#'   total_cores = 1, # Number of BLAST process to start in parallel
 #'   perc_qcov_hsp = 80, # minimum percentage coverage of query sequence by subject sequence cutoff
 #'   num_threads = 1, # number of threads/cores to run each blast on
-#'   total_cores = 8, # number of total threads/cores to allocate all blast searches
 #'   # maximum number of alignments/matches to retrieve results for each query sequence
 #'   num_alignments = 3,
 #'   blast_type = "blastn" # blast search engine to use
@@ -30,30 +43,35 @@
 #' }
 #' @export
 parallel_blast <- function(
-    asvs,
-    db_path,
-    out_file = NULL,
-    out_RDS = NULL,
-    num_threads = 1L,
-    blast_type = "blastn",
-    total_cores = 1L,
-    perc_id = 80L,
-    perc_qcov_hsp = 80L,
-    num_alignments = 4L,
-    verbose = FALSE,
-    env_name = "blastr-blast-env") {
+  asvs,
+  db_path,
+  ...,
+  out_file = NULL,
+  out_RDS = NULL,
+  total_cores = 1L,
+  num_threads = 1L,
+  blast_type = "blastn",
+  perc_id = 80L,
+  perc_qcov_hsp = 80L,
+  num_alignments = 4L,
+  verbose = "silent",
+  env_name = "blastr-blast-env"
+) {
   rlang::check_required(asvs)
   rlang::check_required(db_path)
+  rlang::check_dots_empty()
 
   check_cmd(blast_type, env_name = env_name, verbose = verbose)
 
   parallel_set <- FALSE
-
   if (
-    isTRUE(total_cores > 1L) &&
-      isTRUE(mirai::status(.compute = "blastr-cpu")$connections < total_cores)
+    isTRUE(length(asvs) > 1L) &&
+      isTRUE(total_cores > 1L) &&
+      isTRUE(mirai::status()$connections < total_cores)
   ) {
-    mirai::daemons(n = total_cores, .compute = "blastr-cpu")
+    # TODO: @luciorq add withr defer instead of stopping daemons at the end
+    mirai::daemons(n = total_cores)
+    # NOTE: @luciorq only set to true if daemons were created by this function
     parallel_set <- TRUE
   }
 
@@ -73,7 +91,7 @@ parallel_blast <- function(
           env_name = env_name
         )
       },
-      num_threads = 1L,
+      num_threads = num_threads,
       blast_type = blast_type,
       num_alignments = num_alignments,
       db_path = db_path,
@@ -91,7 +109,7 @@ parallel_blast <- function(
     class(blast_res) <- c("tbl_df", "tbl", "data.frame")
   }
 
-  if (isTRUE(!is.na(out_file) && !is.null(out_file))) {
+  if (!rlang::is_na(out_file) && !rlang::is_null(out_file)) {
     readr::write_csv(
       x = blast_res,
       file = out_file,
@@ -99,7 +117,7 @@ parallel_blast <- function(
     )
   }
 
-  if (isTRUE(!is.na(out_RDS) && !is.null(out_RDS))) {
+  if (!rlang::is_na(out_RDS) && !rlang::is_null(out_RDS)) {
     readr::write_rds(
       x = blast_res,
       file = out_RDS
@@ -107,7 +125,7 @@ parallel_blast <- function(
   }
 
   if (isTRUE(total_cores > 1L) && isTRUE(parallel_set)) {
-    mirai::daemons(n = 0L, .compute = "blastr-cpu")
+    mirai::daemons(n = 0L)
   }
 
   return(blast_res)