+++ tests/testthat/test-get_tax_id.R
@@ -2,7 +2,7 @@ testthat::test_that("Get single taxid", {
   testthat::skip_on_cran()
   testthat::skip_if_offline()
 
-  res <- get_tax_by_taxID(organisms_taxIDs = "63221", verbose = FALSE)
+  res <- get_tax_by_taxID(organisms_taxIDs = "63221", verbose = "silent")
 
   testthat::expect_s3_class(res, "tbl_df")
 
@@ -16,7 +16,7 @@ testthat::test_that("Get single taxid", {
 testthat::test_that("Get single taxid empty df", {
   testthat::skip_on_cran()
   testthat::skip_if_offline()
-  res <- get_tax_by_taxID(organisms_taxIDs = "10000000", verbose = FALSE)
+  res <- get_tax_by_taxID(organisms_taxIDs = "10000000", verbose = "silent")
 
   testthat::expect_s3_class(res, "tbl_df")
 
@@ -28,7 +28,7 @@ testthat::test_that("Get single taxid empty df", {
 testthat::test_that("Get parallel Single taxid", {
   testthat::skip_on_cran()
   testthat::skip_if_offline()
-  res <- parallel_get_tax(organisms_taxIDs = "63221", verbose = FALSE)
+  res <- parallel_get_tax(organisms_taxIDs = "63221", verbose = "silent")
 
   testthat::expect_equal(res$query_taxID, "63221")
 
@@ -42,7 +42,7 @@ testthat::test_that("Get parallel Multiple taxid", {
     organisms_taxIDs = c("63221", "10000000"),
     total_cores = 1,
     retry_times = 2,
-    verbose = FALSE
+    verbose = "silent"
   )
 
   testthat::expect_equal(res$query_taxID, "63221")
@@ -58,7 +58,7 @@ testthat::test_that("Get parallel single-thread Multiple taxid", {
     organisms_taxIDs = c("63221", "10000000"),
     total_cores = 2,
     retry_times = 2,
-    verbose = FALSE
+    verbose = "silent"
   )
 
   testthat::expect_equal(res$query_taxID, "63221")
@@ -74,7 +74,7 @@ testthat::test_that("`parallel_get_tax()` with `retry_times = 0` ", {
     organisms_taxIDs = c("63221", "10000000"),
     total_cores = 1,
     retry_times = 0,
-    verbose = FALSE
+    verbose = "silent"
   )
 
   testthat::expect_equal(res$query_taxID, "63221")