+++ tests/testthat/test-run_blast.R
@@ -1,7 +1,12 @@
 testthat::test_that("parallel_blast works", {
   testthat::skip_on_cran()
   testthat::skip_if_offline()
-  ASVs_test <- readLines(fs::path_package("BLASTr", "extdata", "asvs_test", ext = "txt"))
+  ASVs_test <- readLines(fs::path_package(
+    "BLASTr",
+    "extdata",
+    "asvs_test",
+    ext = "txt"
+  ))
   db_path <- tmp_blast_db_path
 
   blast_res <- parallel_blast(
@@ -15,7 +20,7 @@ testthat::test_that("parallel_blast works", {
     perc_qcov_hsp = 80,
     num_alignments = 4,
     blast_type = "blastn",
-    verbose = FALSE
+    verbose = "silent"
   )
 
   testthat::expect_s3_class(blast_res, "tbl_df")
@@ -25,7 +30,12 @@ testthat::test_that("run_blast works", {
   testthat::skip_on_cran()
   testthat::skip_if_offline()
 
-  ASVs_test <- readLines(fs::path_package("BLASTr", "extdata", "asvs_test", ext = "txt"))
+  ASVs_test <- readLines(fs::path_package(
+    "BLASTr",
+    "extdata",
+    "asvs_test",
+    ext = "txt"
+  ))
   db_path <- tmp_blast_db_path
 
   blast_res <- run_blast(
@@ -41,7 +51,12 @@ testthat::test_that("run_blast fails", {
   testthat::skip_on_cran()
   testthat::skip_if_offline()
 
-  ASVs_test <- readLines(fs::path_package("BLASTr", "extdata", "asvs_test", ext = "txt"))
+  ASVs_test <- readLines(fs::path_package(
+    "BLASTr",
+    "extdata",
+    "asvs_test",
+    ext = "txt"
+  ))
   # `tmp_blast_db_path <- fs::file_temp("minimal_db_blast_")`
   db_path <- tmp_blast_db_path
 
@@ -50,7 +65,7 @@ testthat::test_that("run_blast fails", {
       asv = ASVs_test[1],
       db_path = db_path,
       perc_id = "MISSING_VALUE_STRInG",
-      verbose = FALSE
+      verbose = "silent"
     ),
     class = "blastr_error_blast_run"
   )